requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.1.7
django-cache-machine==1.2.0
django-redis==5.2.0
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
Pillow==9.4.0
pymemcache==4.0.0
python-memcached==1.59
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
